requirements.txt
gitpython>=3.1.0
PyGitHub>=1.55.0
requests>=2.25.0
ruff>=0.1.0
black>=22.0.0
isort>=5.10.0